requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
